+++ test/src/Test/Php82/TestClassReadonly.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Eloquent\Phony\Test\Php82;
+
+readonly class TestClassReadonly
+{
+  public string $propertyA;
+  public int $propertyB;
+
+  public function __construct() {
+    $this->propertyA = 'a';
+    $this->propertyB = 111;
+  }
+}